@@ -10,7 +10,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=3.0.3 <4.0.0"
 
 dependencies:
   flutter:
@@ -18,12 +18,12 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
+  cupertino_icons: ^1.0.5
   flutter_icons:
-#    path: ../
-     git:
-       url: https://github.com/flutter-studio/flutter-icons.git
-       ref: 1.1.0
+   path: ../
+    # git:
+     #  url: https://github.com/flutter-studio/flutter-icons.git
+     #  ref: 1.1.0
 dev_dependencies:
   flutter_test:
     sdk: flutter
