@@ -1,6 +1,6 @@
+import 'dart:math' as math;
 
 import 'package:flutter/material.dart';
-import 'dart:math' as math;
 
 Widget _defaultTransitionBuilder(Widget child, Animation<double> animation) =>
     ScaleTransition(
@@ -25,18 +25,18 @@ class IconToggle extends StatefulWidget {
   final Color activeColor;
   final Color inactiveColor;
   final bool value;
-  final ValueChanged<bool> onChanged;
+  final ValueChanged<bool>? onChanged;
   final AnimatedSwitcherTransitionBuilder transitionBuilder;
   final Duration duration;
-  final Duration reverseDuration;
+  final Duration? reverseDuration;
   @override
   _IconToggleState createState() => _IconToggleState();
 }
 
 class _IconToggleState extends State<IconToggle>
     with SingleTickerProviderStateMixin {
-  AnimationController _controller;
-  Animation<double> _position;
+  late AnimationController _controller;
+  late Animation<double> _position;
   bool _cancel = false;
 
   @override
@@ -51,14 +51,14 @@ class _IconToggleState extends State<IconToggle>
       if (status == AnimationStatus.dismissed &&
           widget.onChanged != null &&
           _cancel == false) {
-        widget.onChanged(!widget.value);
+        widget.onChanged?.call(!widget.value);
       }
     });
   }
 
   @override
   void dispose() {
-    _controller?.dispose();
+    _controller.dispose();
     super.dispose();
   }
 
@@ -68,14 +68,14 @@ class _IconToggleState extends State<IconToggle>
       behavior: HitTestBehavior.opaque,
       onTapDown: (event) {
         _cancel = false;
-        _controller?.forward();
+        _controller.forward();
       },
       onTapUp: (event) {
-        _controller?.reverse();
+        _controller.reverse();
       },
       onTapCancel: () {
         _cancel = true;
-        _controller?.reverse();
+        _controller.reverse();
       },
       child: Padding(
         padding: const EdgeInsets.all(10.0),
@@ -88,7 +88,9 @@ class _IconToggleState extends State<IconToggle>
             reverseDuration: widget.reverseDuration,
             transitionBuilder: widget.transitionBuilder,
             child: Icon(
-              widget.value ? widget.selectedIconData : widget.unselectedIconData,
+              widget.value
+                  ? widget.selectedIconData
+                  : widget.unselectedIconData,
               color: widget.value ? widget.activeColor : widget.inactiveColor,
               size: 22,
               key: ValueKey<bool>(widget.value),
@@ -102,19 +104,21 @@ class _IconToggleState extends State<IconToggle>
 
 class _IconToggleable<T> extends AnimatedWidget {
   _IconToggleable({
-    Animation<T> listenable,
-    this.activeColor,
-    this.inactiveColor,
-    this.child,
+    required Animation<double> listenable,
+    required this.activeColor,
+    required this.inactiveColor,
+    required this.child,
   }) : super(listenable: listenable);
   final Color activeColor;
   final Color inactiveColor;
   final Widget child;
+
+  Animation<double> get animation => listenable as Animation<double>;
   @override
   Widget build(BuildContext context) {
     return CustomPaint(
       painter: _IconPainter(
-        position: listenable,
+        position: animation,
         activeColor: activeColor,
         inactiveColor: inactiveColor,
       ),
@@ -125,21 +129,27 @@ class _IconToggleable<T> extends AnimatedWidget {
 
 class _IconPainter extends CustomPainter {
   _IconPainter({
-    @required this.position,
-    this.activeColor,
-    this.inactiveColor,
+    required this.position,
+    required this.activeColor,
+    required this.inactiveColor,
   });
   final Animation<double> position;
   final Color activeColor;
   final Color inactiveColor;
 
-  double get _value => position != null ? position.value : 0;
+  double get _value {
+    final localPosition = position;
+    return localPosition != null ? localPosition.value : 0;
+  }
 
   @override
   void paint(Canvas canvas, Size size) {
+    final color = Color.lerp(inactiveColor, activeColor, _value)
+            ?.withOpacity(math.min(_value, 0.15)) ??
+        activeColor;
+
     final Paint paint = Paint()
-      ..color = Color.lerp(inactiveColor, activeColor, _value)
-          .withOpacity(math.min(_value, 0.15))
+      ..color = color
       ..style = PaintingStyle.fill
       ..strokeWidth = 2.0;
     canvas.drawCircle(
